+++ icevision/models/rcnn/mask_rcnn/model.py
@@ -14,6 +14,7 @@ def model(
     num_classes: int,
     backbone: Optional[nn.Module] = None,
     remove_internal_transforms: bool = True,
+    pretrained: bool = True,
     **mask_rcnn_kwargs
 ) -> nn.Module:
     """MaskRCNN model implemented by torchvision.
@@ -24,6 +25,9 @@ def model(
         remove_internal_transforms: The torchvision model internally applies transforms
         like resizing and normalization, but we already do this at the `Dataset` level,
         so it's safe to remove those internal transforms.
+        pretrained: Argument passed to `maskrcnn_resnet50_fpn` if `backbone is None`.
+        By default it is set to True: this is generally used when training a new model (transfer learning).
+        `pretrained = False`  is used during inference (prediction) for cases where the users have their own pretrained weights.
         **mask_rcnn_kwargs: Keyword arguments that internally are going to be passed to
         `torchvision.models.detection.mask_rcnn.MaskRCNN`.
 
@@ -31,7 +35,7 @@ def model(
         A Pytorch `nn.Module`.
     """
     if backbone is None:
-        model = maskrcnn_resnet50_fpn(pretrained=True, **mask_rcnn_kwargs)
+        model = maskrcnn_resnet50_fpn(pretrained=pretrained, **mask_rcnn_kwargs)
 
         in_features_box = model.roi_heads.box_predictor.cls_score.in_features
         model.roi_heads.box_predictor = FastRCNNPredictor(in_features_box, num_classes)